src/ingest/account_mapper.py
"""Account mapping from ERP codes to financial categories."""

from pathlib import Path
from typing import Dict, List, Tuple, Optional
from decimal import Decimal
import yaml

from src.models import AccountEntry, MappedData


class AccountMapper:
    """
    Maps ERP account codes to financial categories (Chapter 1-4 structure).
    
    Uses configuration file to define mappings:
    - Chapter 1: Revenue, COGS, OpEx, EBITDA, Depreciation, EBIT, Financial, Taxes
    - Chapter 2: AR, Inventory, AP
    - Chapter 3: PPE, Intangibles, Other Capital
    - Chapter 4: Cash, Debt, Equity
    
    Attributes:
        config_path: Path to mapping configuration file
        mapping: Dictionary of account code → category mappings
        validation_rules: Rules for cross-account validation
    """
    
    def __init__(self, config_path: str) -> None:
        """
        Initialize account mapper.
        
        Args:
            config_path: Path to YAML mapping configuration file
            
        TODO: Load and parse configuration file
        TODO: Validate configuration structure
        TODO: Initialize mapping and validation rules
        """
        self.config_path = Path(config_path)
        self.mapping: Dict[str, str] = {}
        self.validation_rules: Dict[str, any] = {}
    
    def load_config(self) -> None:
        """
        Load mapping configuration from YAML file.
        
        Configuration structure:
        ```yaml
        chapters:
          chapter_1:
            gross_revenue: ["4100", "4110", "4120"]
            returns_deductions: ["4900", "4910"]
            cogs: ["5100", "5110"]
            ...
          chapter_2:
            accounts_receivable: ["1200", "1210"]
            ...
          chapter_3:
            ppe_gross: ["2100", "2110"]
            ...
          chapter_4:
            cash: ["1010", "1020"]
            short_term_debt: ["2100"]
            long_term_debt: ["2200"]
            equity: ["3100", "3200"]
        ```
        
        Raises:
            FileNotFoundError: If config file doesn't exist
            yaml.YAMLError: If config is invalid
            
        TODO: Load and parse YAML configuration
        TODO: Build internal mapping dictionary
        TODO: Validate all required chapters present
        """
        pass
    
    def map_accounts(
        self,
        accounts: List[AccountEntry],
        period: str
    ) -> MappedData:
        """
        Map raw account entries to financial categories.
        
        Aggregates accounts by their configured categories and creates MappedData
        with all Chapter 1-4 financial inputs.
        
        Args:
            accounts: List of raw account entries from ERP
            period: Period identifier (e.g., "202401")
        
        Returns:
            MappedData: Mapped and aggregated financial data
            
        Raises:
            ValueError: If required accounts missing or unmapped accounts found
            
        TODO: Filter accounts by period
        TODO: Aggregate by category using mapping
        TODO: Detect unmapped accounts and log warnings
        TODO: Calculate derived values (gross profit, EBITDA, EBIT, etc)
        TODO: Detect company from account metadata
        TODO: Return complete MappedData object
        TODO: Handle missing optional accounts gracefully
        """
        pass
    
    def validate_mapping(self, mapped: MappedData) -> Tuple[bool, List[str]]:
        """
        Validate mapped data against business rules.
        
        Validation checks:
        - Revenue components make sense (returns < revenue)
        - COGS < Revenue
        - Expenses are reasonable (OpEx < Revenue)
        - Balance sheet consistency (AR + Inv should be positive)
        - Working capital components reasonable
        - Debt should be positive
        - Equity components consistent
        
        Args:
            mapped: Mapped financial data to validate
        
        Returns:
            Tuple[bool, List[str]]: (is_valid, list of validation messages/warnings)
            
        TODO: Implement business rule validation
        TODO: Check for unrealistic values
        TODO: Validate relationships between accounts
        TODO: Generate descriptive validation messages
        TODO: Return warnings (not just errors)
        """
        pass
