pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cashflow-story-pipeline"
version = "1.0.0"
description = "Automated CashFlow Story analysis pipeline — ERP XML to board-ready insights"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Rodrigo Aquino", email = "rodrigotaquino@gmail.com" },
]
keywords = ["cashflow", "financial-analysis", "erp", "pipeline", "ai"]

dependencies = [
    "pydantic>=2.0,<3.0",
    "pandas>=2.0,<3.0",
    "lxml>=5.0",
    "openpyxl>=3.1,<4.0",
    "jinja2>=3.1,<4.0",
    "matplotlib>=3.8",
    "weasyprint>=60.0",
    "anthropic>=0.40",
    "pyyaml>=6.0",
    "click>=8.1,<9.0",
    "structlog>=24.0",
    "watchdog>=4.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-mock>=3.12",
    "ruff>=0.4",
    "mypy>=1.10",
    "pre-commit>=3.7",
]

[project.scripts]
cashflow = "src.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow",
    "integration: marks integration tests requiring external services",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "ANN"]
ignore = ["ANN101", "ANN102"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]

[tool.coverage.report]
fail_under = 90
show_missing = true
